License change and version number update
I don't think I would be able to contact the other members of my group, and since I more or less unilaterally licensed this program under GPLv3 in the first place, I'm going to go ahead and change it to a 2-clause BSD license. It's my understanding that you can't use a GPL program with an MSPL library like CompareNetObjects.